tamuhack25/llama3/requirements.txt
boto3
awscli
newspaper3k
lxml_html_clean